chore: modify docker container setup to use SuiteCRM files on the host
This commit modifies the local development docker container setup to work
with SuiteCRM files on the host.  This allows us to work with a SuiteCRM
distribution, develop those files locally on the host, and commit them to
version control.  The commit includes a new Dockerfile built from a PHP
base that is set up to receive the SuiteCRM files from the host, there is
a configuration file for SuiteCRM to silently install itself, and a small
patch to the PHP configuration.  The compose file is adjusted to remove a
lot of configuration and to share the locally hosted files with the
container.